+++ controllers/contacts/index.js
@@ -1,0 +1,15 @@
+const listContacts = require("./listContacts");
+const getContactById = require("./getContactById");
+const removeContact = require("./removeContact");
+const addContact = require("./addContact");
+const updateContact = require("./updateContact");
+const updateStatusContact = require("./updateStatusContact");
+
+module.exports = {
+    listContacts,
+    getContactById,
+    removeContact,
+    addContact,
+    updateContact,
+    updateStatusContact
+}
