+++ controllers/contacts/listContacts.js
@@ -1,0 +1,10 @@
+const {Contact} = require("../../models/contact")
+
+const {ControllerWrapper} = require("../../utils/index");
+
+const listContacts = async (req, res) => {
+  const result = await Contact.find();
+  res.json(result);    
+};
+
+module.exports = ControllerWrapper(listContacts);
